Add git clone and mail support

+++ deploy_images.py
@@ -7,21 +7,67 @@ Wrapper for Docker related utils.
 
 
 import docker
+import getpass
 import os
+import smtplib
+import ssl
 import subprocess
 import sys
 import time
 
 
 NO_OF_CONTAINERS = 2
+GIT_URI = 'https://github.com/JonnieDoe/Deploy_web_app.git'
+REPO_NAME = 'Deploy_web_app'
+# Mail settings
+SMTP_SERVER = "smtp.gmail.com"
+PORT = 587  # For starttls
+
+
+def send_mail():
+    """Send mail to signal the containers are up and running."""
+    sender_email = input("Sender mail address:  ")
+    password = getpass.getpass("Type your password and press enter:  ")
+    receiver_email = input("Receiver mail address:  ")
+    message = """\
+    Subject: Docker status
+
+    Docker containers are up and running.."""
+
+    context = ssl.create_default_context()
+    with smtplib.SMTP(SMTP_SERVER, PORT) as server:
+        server.ehlo()
+        server.starttls(context=context)
+        server.ehlo()
+        server.login(sender_email, password)
+        server.sendmail(sender_email, receiver_email, message)
 
 
 def main():
     """Main thread."""
 
-    cwd = sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+    parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+    cwd = sys.path.append(parent_dir)
 
-    cmds_to_execute = ["docker-compose down -v ", "docker-compose up --detach --force-recreate --build"]
+    # Fetch the remote repo
+    cmd = 'git clone {git_uri}'.format(git_uri=GIT_URI)
+    sp = subprocess.Popen(
+        cmd,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        universal_newlines=True,
+        bufsize=1,
+        shell=True,
+        cwd=cwd
+    )
+    sp.wait()
+
+    if sp.returncode != 0:
+        print("Could not execute command: '{cmd}'".format(cmd=cmd))
+        sys.exit(1)
+
+    # Bring up the containers
+    cmds_to_execute = ["docker-compose down -v", "docker-compose up --detach --force-recreate --build"]
     for cmd in cmds_to_execute:
         sp = subprocess.Popen(
             cmd,
@@ -30,7 +76,7 @@ def main():
             universal_newlines=True,
             bufsize=1,
             shell=True,
-            cwd=cwd
+            cwd=os.path.join(parent_dir, REPO_NAME)
         )
         # Poll process.stdout to show stdout live
         while True:
@@ -62,8 +108,15 @@ def main():
 
     list_of_containers = client.containers.list(filters={"status": "running"})
     if len(list_of_containers) < NO_OF_CONTAINERS:
-        print("Not all containers are running!")
+        print("Not all containers are running! Exiting")
         sys.exit(1)
+
+    print("\nAll containers are up and running: {list_of_containers}\n".format(list_of_containers=list_of_containers))
+
+    try:
+        send_mail()
+    except Exception as exec_err:
+        print("Could not send mail notification!\nReason: {reason}".format(reason=exec_err))
 
 
 ####################################################################################################
